Allow non-integral days for watchlists and RC
Fixes bug 19296.  This was a regression due to the preferences work.

+++ includes/Preferences.php
@@ -689,7 +689,7 @@ class Preferences {
 		## RecentChanges #####################################
 		$defaultPreferences['rcdays'] =
 				array(
-					'type' => 'int',
+					'type' => 'float',
 					'label-message' => 'recentchangesdays',
 					'section' => 'rc/display',
 					'min' => 1,
@@ -747,7 +747,7 @@ class Preferences {
 		## Watchlist #####################################
 		$defaultPreferences['watchlistdays'] =
 				array(
-					'type' => 'int',
+					'type' => 'float',
 					'min' => 0,
 					'max' => 7,
 					'section' => 'watchlist/display',
